// client/src/store/dialogs/actionTypes.ts
export enum ActionTypes {
  FETCH_DIALOGS_START = "DIALOGS:FETCH_DIALOGS_START",
  FETCH_DIALOGS_FAIL = "DIALOGS:FETCH_DIALOGS_FAIL",
  FETCH_DIALOGS_SUCCESS = "DIALOGS:FETCH_DIALOGS_SUCCESS",
  SET_CURRENT_DIALOG = "DIALOGS:SET_CURRENT_DIALOG",
  ADD_DIALOG = "DIALOGS:ADD_DIALOG",
  POST_DIALOG_START = "DIALOGS:POST_DIALOG_START",
  POST_DIALOG_FAIL = "DIALOGS:POST_DIALOG_FAIL",
  FILTER = "DIALOGS:FILTER",
}